requirements.txt
uvicorn
fastapi
fastmcp
pydantic-settings
httpx
keyring>=25.2
pycryptodome
python-jose
requests